refactor: change executor implementation

+++ src/runtime/globals.rs
@@ -8,7 +8,7 @@ thread_local! {
 
 thread_local! {
     /// This is the task thread local. It determines which task is currently being executed.
-    pub(crate) static TASK_ID: Cell<Option<usize>> = Cell::new(None);
+    pub(crate) static TASK_ID: Cell<Option<u64>> = Cell::new(None);
 }
 
 /// Returns a handle to the currently running [`Runtime`].
